server/requirements.txt
fastapi
uvicorn[standard]
cryptography
pillow
numpy
opencv-python
pytest
python-multipart
